Initial Commit, until Comparing Shots P.717
The camera in the top left shows the list of shots, you can scroll
through them with Q and E.

you can remove all the shots by going to _TargetCamera and checking
"Check to delete player prefs"

You take shots with left click and replace shots with right click.

You can save the list of current shots with leftShift+S

Every time you take a shot it records the position and rotation and a
raycast.

It's a pretty neat game.